Update metodo calcolo gintonic

+++ src/main/java/com/aiplus/aiplus/entities/stockentities/GinTonic.java
@@ -51,5 +51,5 @@ public class GinTonic {
     @OneToMany(mappedBy = "ginTonic", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<GarnishQuantity> garnishes;
 
-    private int finalPrice;
+    private double finalPrice;
 }
